chatbot/requirements.txt
chainlit==1.3.2
pydantic==2.9.2
llama-index==0.12.3
llama-index-llms-openai==0.3.2
llama-index-agent-openai==0.4.0
python-dotenv==1.0.1